Change Route structure for auth route group
