+++ src/app/(auth)/[...not-found]/page.tsx
@@ -1,4 +1,10 @@
+import { Metadata } from 'next';
 import { notFound } from 'next/navigation';
+
+export const metadata: Metadata = {
+  title: 'Page not found',
+  description: 'Page not found',
+};
 
 export default function NotFound() {
   notFound();
