+++ src/app/(auth)/auth/login/page.tsx
@@ -1,18 +1,19 @@
 import type { Metadata } from 'next';
 import AuthForm from '@/_components/AuthForm';
-import SignUpForm from '@/_components/SignUpForm';
+import LoginForm from '@/_components/LoginForm';
 import AuthLink from '@/_components/AuthLink';
 
 export const metadata: Metadata = {
-  title: 'Sign Up',
+  title: 'Sign In',
 };
 
-export default function SignUpPage() {
+export default function LoginPage() {
   return (
-    <AuthForm heading='Create an account'>
-      <SignUpForm />
+    <AuthForm heading='Login to account'>
+      <LoginForm />
       <p className='text-xs mt-3 text-slate-400'>
-        Already have an account? <AuthLink href='/login'>Sign In</AuthLink>
+        Don&apos;t have an account?{' '}
+        <AuthLink href='/auth/signup'>Sign Up</AuthLink>
       </p>
     </AuthForm>
   );
